byte conversions

+++ ds/src/main/java/cs451/Host.java
@@ -7,7 +7,7 @@ public class Host {
 
     private static final String IP_START_REGEX = "/";
 
-    private int id;
+    private byte id;
     private String ip;
     private int port = -1;
 
@@ -15,7 +15,7 @@ public class Host {
 
     public boolean populate(String idString, String ipString, String portString) {
         try {
-            id = Integer.parseInt(idString);
+            id = (byte) Integer.parseInt(idString);
 
             String ipTest = InetAddress.getByName(ipString).toString();
             if (ipTest.startsWith(IP_START_REGEX)) {
@@ -43,7 +43,7 @@ public class Host {
         return true;
     }
 
-    public int getId() {
+    public byte getId() {
         return id;
     }
