+++ ds/src/main/java/cs451/commonUtils/CommonUtils.java
@@ -51,7 +51,7 @@ public class CommonUtils {
             int to = Integer.parseInt(contents[1]);
 
             for(int i=0; i<repetitions; i++){
-                messageQueue.add(new Message(processID, to, id++));
+                messageQueue.add(new Message((byte) processID, (byte) to, id++));
             }
 
         }
