+++ ds/src/main/java/cs451/messaging/MessageBatch.java
@@ -13,7 +13,7 @@ public class MessageBatch {
     public static void broadcastBatch(int numberOfMessages, cs451.structures.Process process){
         ArrayList<Message>batch = new ArrayList<>();
         for(int i=0; i<numberOfMessages; i++){
-            batch.add(new Message(process.getId(), -1, i+1));
+            batch.add(new Message((byte) process.getId(), (byte) -1, i+1));
             if(batch.size() == MESSAGES_PER_BATCH){
                 //System.out.println("Batch:" + batch);
                 process.broadcastBatch(batch);
@@ -32,7 +32,7 @@ public class MessageBatch {
 
         ArrayList<Message>batch = new ArrayList<>();
         for(int i=0; i<numberOfMessages; i++){
-            batch.add(new Message(process.getId(), toHost.getId(), i+1));
+            batch.add(new Message((byte) process.getId(), (byte) toHost.getId(), i+1));
             if(batch.size() == MESSAGES_PER_BATCH){
                 //System.out.println("Batch:" + batch);
                 process.sendBatch(batch, toHost);
